contents of LTVT_Source_Code_v0_19_6.zip from Nov 28, 2008
Two zip archives were created on Nov 28, 2008, this one
and the one for version v0_19_07.

No stand-alone executable is provided for either, they
skip directly from LTVT_v0_19_5.exe on Nov 20, 2008 to
LTVT_v0_19_7_1.exe on NOV 28, 2008.

These dates are upload dates, so if you care about that
sort of thing you should be comparing the changes to the
dates of the source files and especially their content.